Version 15 (14 failed)

+++ backend/foodgram/users/views.py
@@ -113,10 +113,15 @@ class CustomUserViewSet(UserViewSet):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         if request.method == 'DELETE':
             subscription = Subscription.objects.filter(
-                user=subscriber, author=author
+                subscriber=subscriber, author=author
             )
-            subscription.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if subscription.exists():
+                subscription.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(
+                {'error': 'Вы не подписаны на этого пользователя.'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
     @action(
         detail=False, methods=('GET',),
